Prototype Version
Some features hard coded, but added textures, animations for players and for 1 enemy. There is AI in place for the voidling, however it is rather basic and inefficient. Improvements are not top priority. Special overlay is in place for unchanging animations such as water and lava on the ground. Full implementation is soon to come, will improve performance via sprite-sheet in the near future, also planned is reading map from text files + map editor.